requirements.txt
fastapi
uvicorn
psycopg2-binary
ulid-py
pydantic[email]